fix: refactor to custom hook
